Extract zip download logic into ZipDownloadConcern
Moves EXTENSIONS constant, build_zip, write_zip_entry, and ext_for out of
ClassroomKitsController into a shared concern to eliminate duplication.
